Don't log an error on unauthorized requests (#170)
* Don't log an error on unauthorized requests

When trying to fetch the user's session, if the error hit was due to a 401 Unauthorized response from Sirius, log an info-level message rather than an error-level one. This prevents false positives in alarms, because 401 responses from Sirius are BAU behaviour.

#patch

* Use `apierror.Unauthorized` in Authenticate

So that it can be tested with `errors.Is()` rather than using string-matching

#patch